backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
fastapi==0.104.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9

# FIXED → Supabase needs httpx + http2 and version <0.29
httpx[http2]>=0.26,<0.29

hyperframe==6.1.0
idna==3.11
multidict==6.7.0
numpy==2.2.6
opencv-python-headless>=4.6.0
packaging==25.0
pillow==10.1.0

# YOLO Support (OPTIONAL - disabled by default)
# ultralytics requires PyTorch which adds ~5GB with CUDA
# Only enable if you need YOLO signature/face detection
# To enable: Set YOLO_SIGNATURE_ENABLED=true or YOLO_FACE_ENABLED=true
# and run: pip install ultralytics --index-url https://download.pytorch.org/whl/cpu
ultralytics>=8.0.0

# Supabase ecosystem (compatible with httpx>=0.26,<0.29)
postgrest==2.24.0
storage3==2.24.0
realtime==2.24.0
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0

propcache==0.4.1
pycparser==2.23
pydantic-core==2.41.5
pydantic-settings==2.1.0
pydantic==2.12.4
pymupdf==1.23.14
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.3
requests==2.31.0
sniffio==1.3.1
starlette==0.27.0
strenum==0.4.15
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.24.0
uvloop==0.22.1 ; platform_system != "Windows"
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
langsmith==0.1.129
google-generativeai>=0.8.0